+++ src/ts/EventHandlers/MRAIDEventHandler.ts
@@ -15,6 +15,8 @@ import { RequestError } from 'Errors/RequestError';
 import { DiagnosticError } from 'Errors/DiagnosticError';
 import { FinishState } from 'Constants/FinishState';
 import { Placement } from 'Models/Placement';
+import { Configuration } from 'Models/Configuration';
+import { GDPREventAction, GdprManager } from 'Managers/GdprManager';
 
 export class MRAIDEventHandler implements IMRAIDViewHandler {
 
@@ -27,6 +29,8 @@ export class MRAIDEventHandler implements IMRAIDViewHandler {
     private _campaign: MRAIDCampaign;
     private _request: Request;
     private _placement: Placement;
+    private _configuration: Configuration;
+    private _gdprManager: GdprManager;
 
     constructor(nativeBridge: NativeBridge, adUnit: MRAIDAdUnit, parameters: IMRAIDAdUnitParameters) {
         this._nativeBridge = nativeBridge;
@@ -38,6 +42,8 @@ export class MRAIDEventHandler implements IMRAIDViewHandler {
         this._campaign = parameters.campaign;
         this._placement = parameters.placement;
         this._request = parameters.request;
+        this._configuration = parameters.configuration;
+        this._gdprManager = parameters.gdprManager;
     }
 
     public onMraidClick(url: string): Promise<void> {
@@ -117,6 +123,13 @@ export class MRAIDEventHandler implements IMRAIDViewHandler {
             this._adUnit.getMRAIDView().hide();
             endScreen.show();
         }
+    }
+
+    public onGDPRPopupSkipped(): void {
+        if (!this._configuration.isOptOutRecorded()) {
+            this._configuration.setOptOutRecorded(true);
+        }
+        this._gdprManager.sendGDPREvent(GDPREventAction.SKIP);
     }
 
     private handleClickAttribution() {