+++ src/ts/EventHandlers/PromoEventHandler.ts
@@ -3,7 +3,7 @@ import { PurchasingUtilities, IPromoPayload, IPromoRequest } from 'Utilities/Pur
 import { Configuration } from 'Models/Configuration';
 import { FinishState } from 'Constants/FinishState';
 import { Placement } from 'Models/Placement';
-import { OperativeEventManager } from 'Managers/OperativeEventManager';
+import { GdprManager, GDPREventAction } from 'Managers/GdprManager';
 
 export class PromoEventHandler {
 
@@ -35,10 +35,10 @@ export class PromoEventHandler {
         PurchasingUtilities.sendPromoPayload(JSON.stringify(iapPayload));
     }
 
-    public static onGDPRPopupSkipped(configuration: Configuration, operativeEventManager: OperativeEventManager): void {
+    public static onGDPRPopupSkipped(configuration: Configuration, gdprManager: GdprManager): void {
         if (!configuration.isOptOutRecorded()) {
             configuration.setOptOutRecorded(true);
-            operativeEventManager.sendGDPREvent('skip');
+            gdprManager.sendGDPREvent(GDPREventAction.SKIP);
         }
     }
 }