+++ src/ts/Managers/OperativeEventManager.ts
@@ -31,11 +31,6 @@ export interface IOperativeEventManagerParams<T extends Campaign> {
     campaign: T;
 }
 
-export enum GDPREventSource {
-    METADATA = 'metadata',
-    USER = 'user'
-}
-
 export class OperativeEventManager {
 
     public static setTestBaseUrl(baseUrl: string): void {
@@ -61,25 +56,6 @@ export class OperativeEventManager {
 
     public static getPreviousPlacementId(): string | undefined {
         return OperativeEventManager.PreviousPlacementId;
-    }
-
-    public static sendGDPREvent(action: string, deviceInfo: DeviceInfo, clientInfo: ClientInfo, configuration: Configuration, source?: GDPREventSource): Promise<void> {
-        let infoJson: any = {
-            'adid': deviceInfo.getAdvertisingIdentifier(),
-            'action': action,
-            'projectId': configuration.getUnityProjectId(),
-            'platform': Platform[clientInfo.getPlatform()].toLowerCase(),
-            'gameId': clientInfo.getGameId()
-        };
-        if (source) {
-            infoJson = {
-                ... infoJson,
-                'source': source
-            };
-        }
-
-        HttpKafka.sendEvent('ads.events.optout.v1.json', KafkaCommonObjectType.EMPTY, infoJson);
-        return Promise.resolve();
     }
 
     private static VideoEventBaseUrl: string = 'https://adserver.unityads.unity3d.com/mobile/gamers';
@@ -252,10 +228,6 @@ export class OperativeEventManager {
         };
 
         return this.createUniqueEventMetadata(placement, this._sessionManager.getGameSessionId(), this._gamerServerId, OperativeEventManager.getPreviousPlacementId(), videoOrientation, adUnitStyle).then(fulfilled);
-    }
-
-    public sendGDPREvent(action: string, source?: GDPREventSource): Promise<void> {
-        return OperativeEventManager.sendGDPREvent(action, this._deviceInfo, this._clientInfo, this._configuration, source);
     }
 
     public setGamerServerId(serverId: string | undefined): void {