+++ src/ts/Views/Promo.ts
@@ -60,7 +60,7 @@ export class Promo extends View<{}> implements IPrivacyHandler {
             {
                 event: 'click',
                 listener: (event: Event) => this.onPrivacyEvent(event),
-                selector: '.icon-info'
+                selector: '.icon-gdpr'
             }
         ];
 
@@ -88,16 +88,7 @@ export class Promo extends View<{}> implements IPrivacyHandler {
     public show(): void {
         super.show();
         window.addEventListener('message', this._messageHandler);
-
-        if (this._showGDPRBanner && this._privacy instanceof GDPRPrivacy) {
-            this._GDPRPopupElement.style.opacity = '1';
-            this._privacyButtonElement.style.pointerEvents = '1';
-            this._privacyButtonElement.style.visibility = 'hidden';
-        } else {
-            this._GDPRPopupElement.style.pointerEvents = '1';
-            this._GDPRPopupElement.style.visibility = 'hidden';
-            this._iframe!.style.height = '100vh';
-        }
+        this.choosePrivacyShown();
     }
 
     public hide(): void {
@@ -127,6 +118,18 @@ export class Promo extends View<{}> implements IPrivacyHandler {
         // do nothing
     }
 
+    private choosePrivacyShown() {
+        if (!this._gdprPopupClicked && this._showGDPRBanner) {
+            this._GDPRPopupElement.style.visibility = 'visible';
+            this._privacyButtonElement.style.pointerEvents = '1';
+            this._privacyButtonElement.style.visibility = 'hidden';
+        } else {
+            this._privacyButtonElement.style.visibility = 'visible';
+            this._GDPRPopupElement.style.pointerEvents = '1';
+            this._GDPRPopupElement.style.visibility = 'hidden';
+        }
+    }
+
     private onMessage(e: MessageEvent): void {
         const data: any = e.data;
         switch (data.type) {
@@ -154,6 +157,7 @@ export class Promo extends View<{}> implements IPrivacyHandler {
 
         if (this._showGDPRBanner) {
             this._gdprPopupClicked = true;
+            this.choosePrivacyShown();
         }
 
         this._privacy.show();