+++ CHANGELOG.md
@@ -1,4 +1,13 @@
-# Pending
+# Staged
+
+* GDPR banners for all ad units
+
+# Thu Jun  7 17:15:45 UTC 2018
+
+* Fix Purchasing utilities diagnostics
+* Refactor GDPR event sending to GdprManager
+
+# Wed Jun  6 09:37:02 UTC 2018
 
 * Refactor promo purchasing utilities